refactor and more tests
